# Conflicts:
#	app/src/main/java/com/example/myaku_rismu/feature/home/HomeScreen.kt
#	app/src/main/java/com/example/myaku_rismu/feature/home/HomeViewModel.kt
#	app/src/main/java/com/example/myaku_rismu/ui/theme/CustomTheme.kt
#	app/src/main/java/com/example/myaku_rismu/ui/theme/Theme.kt
#	app/src/main/java/com/example/myaku_rismu/ui/theme/Type.kt
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/myaku_rismu/feature/healthDetail/RememberCharRenderData.kt
@@ -15,7 +15,7 @@ import com.patrykandpatrick.vico.core.entry.ChartEntryModelProducer
 import kotlin.math.roundToInt
 
 @Composable
-fun rememberChartRenderData(uiState: HealthDetailState, data: List<Int>): ChartRenderData? {
+fun rememberChartRenderData(uiState: HealthDetailState, data: List<Long>): ChartRenderData? {
     val axisConfig = uiState.axisConfig ?: return null
 
     val modelProducer = remember { ChartEntryModelProducer() }
